+++ components/serverless-page-builder/serverless.js
@@ -1,7 +1,17 @@
+const { join } = require("path");
 const { Component } = require("@serverless/core");
 const { trackComponent } = require("@webiny/tracking");
+const loadJson = require("load-json-file");
 
 const component = "@webiny/serverless-page-builder";
+
+const getDeps = async deps => {
+    const { dependencies } = await loadJson(join(__dirname, "package.json"));
+    return deps.reduce((acc, item) => {
+        acc[item] = dependencies[item];
+        return acc;
+    }, {});
+};
 
 class ServerlessPageBuilder extends Component {
     async default({ track, ...inputs } = {}) {
@@ -13,7 +23,12 @@ class ServerlessPageBuilder extends Component {
 
         // Deploy graphql API
         const apolloService = await this.load("@webiny/serverless-apollo-service");
-        const output = await apolloService({ plugins, ...rest, track: false });
+        const output = await apolloService({
+            plugins,
+            ...rest,
+            track: false,
+            dependencies: await getDeps(["@webiny/api-page-builder"])
+        });
 
         this.state.output = output;
         await this.save();