+++ packages/cli/src/create/template/.prettierrc.js
@@ -1,0 +1,5 @@
+module.exports = {
+    printWidth: 100,
+    parser: "babylon",
+    tabWidth: 4
+};