+++ components/serverless-apollo-service/CHANGELOG.md
@@ -3,6 +3,51 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.6](https://github.com/Webiny/webiny-js/compare/@webiny/serverless-apollo-service@0.1.5...@webiny/serverless-apollo-service@0.1.6) (2019-10-07)
+
+
+### Bug Fixes
+
+* improve handling of boilerplate dependencies. ([cc4404b](https://github.com/Webiny/webiny-js/commit/cc4404b))
+* track proper method on component removal. ([e3cf4b8](https://github.com/Webiny/webiny-js/commit/e3cf4b8))
+
+
+
+
+
+## [0.1.5](https://github.com/Webiny/webiny-js/compare/@webiny/serverless-apollo-service@0.1.4...@webiny/serverless-apollo-service@0.1.5) (2019-10-07)
+
+
+### Bug Fixes
+
+* update aws-api-gateway dependency version ([03f919d](https://github.com/Webiny/webiny-js/commit/03f919d))
+
+
+
+
+
+## [0.1.4](https://github.com/Webiny/webiny-js/compare/@webiny/serverless-apollo-service@0.1.3...@webiny/serverless-apollo-service@0.1.4) (2019-10-07)
+
+
+### Bug Fixes
+
+* add installation of dependencies per component boilerplate. ([647a472](https://github.com/Webiny/webiny-js/commit/647a472))
+
+
+
+
+
+## [0.1.3](https://github.com/Webiny/webiny-js/compare/@webiny/serverless-apollo-service@0.1.2...@webiny/serverless-apollo-service@0.1.3) (2019-10-07)
+
+
+### Bug Fixes
+
+* improve webpack error output. ([e0468cb](https://github.com/Webiny/webiny-js/commit/e0468cb))
+
+
+
+
+
 ## [0.1.2](https://github.com/Webiny/webiny-js/compare/@webiny/serverless-apollo-service@0.1.1...@webiny/serverless-apollo-service@0.1.2) (2019-10-06)