+++ components/serverless-app/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3](https://github.com/Webiny/webiny-js/compare/@webiny/serverless-app@0.1.2...@webiny/serverless-app@0.1.3) (2019-10-07)
+
+**Note:** Version bump only for package @webiny/serverless-app
+
+
+
+
+
+## [0.1.2](https://github.com/Webiny/webiny-js/compare/@webiny/serverless-app@0.1.1...@webiny/serverless-app@0.1.2) (2019-10-07)
+
+**Note:** Version bump only for package @webiny/serverless-app
+
+
+
+
+
 ## [0.1.1](https://github.com/Webiny/webiny-js/compare/@webiny/serverless-app@0.1.0...@webiny/serverless-app@0.1.1) (2019-10-06)