+++ components/serverless-page-builder/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.6](https://github.com/Webiny/webiny-js/compare/@webiny/serverless-page-builder@0.1.5...@webiny/serverless-page-builder@0.1.6) (2019-10-07)
+
+
+### Bug Fixes
+
+* improve handling of boilerplate dependencies. ([cc4404b](https://github.com/Webiny/webiny-js/commit/cc4404b))
+
+
+
+
+
+## [0.1.5](https://github.com/Webiny/webiny-js/compare/@webiny/serverless-page-builder@0.1.4...@webiny/serverless-page-builder@0.1.5) (2019-10-07)
+
+**Note:** Version bump only for package @webiny/serverless-page-builder
+
+
+
+
+
+## [0.1.4](https://github.com/Webiny/webiny-js/compare/@webiny/serverless-page-builder@0.1.3...@webiny/serverless-page-builder@0.1.4) (2019-10-07)
+
+
+### Bug Fixes
+
+* add installation of dependencies per component boilerplate. ([647a472](https://github.com/Webiny/webiny-js/commit/647a472))
+
+
+
+
+
+## [0.1.3](https://github.com/Webiny/webiny-js/compare/@webiny/serverless-page-builder@0.1.2...@webiny/serverless-page-builder@0.1.3) (2019-10-07)
+
+**Note:** Version bump only for package @webiny/serverless-page-builder
+
+
+
+
+
 ## [0.1.2](https://github.com/Webiny/webiny-js/compare/@webiny/serverless-page-builder@0.1.1...@webiny/serverless-page-builder@0.1.2) (2019-10-06)
 
 **Note:** Version bump only for package @webiny/serverless-page-builder