+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1](https://github.com/webiny/webiny-js/compare/@webiny/cli@0.2.0...@webiny/cli@0.2.1) (2019-10-07)
+
+**Note:** Version bump only for package @webiny/cli
+
+
+
+
+
+# [0.2.0](https://github.com/webiny/webiny-js/compare/@webiny/cli@0.1.1...@webiny/cli@0.2.0) (2019-10-07)
+
+
+### Features
+
+* add commands to enable/disable stats tracking. ([62a710d](https://github.com/webiny/webiny-js/commit/62a710d))
+* add webiny config creation. ([d918fc2](https://github.com/webiny/webiny-js/commit/d918fc2))
+
+
+
+
+
 ## [0.1.1](https://github.com/webiny/webiny-js/compare/@webiny/cli@0.1.0...@webiny/cli@0.1.1) (2019-10-06)
 
 **Note:** Version bump only for package @webiny/cli